+++ lbrynet/tests/unit/lbryfilemanager/test_EncryptedFileManager.py
@@ -3,7 +3,7 @@ from twisted.trial import unittest
 from lbrynet import conf
 from lbrynet.file_manager.EncryptedFileDownloader import ManagedEncryptedFileDownloader
 from lbrynet.file_manager.EncryptedFileManager import EncryptedFileManager
-from tests.util import random_lbry_hash
+from lbrynet.tests.util import random_lbry_hash
 
 class TestEncryptedFileManager(unittest.TestCase):